Fixes #5973: Rename github_cache.py to github_cache_loop.py for naming consistency (#6079)
Rename the file to follow the *_loop.py convention used by all other
BaseBackgroundLoop subclasses. Update all importers and remove the
special-case in test_loop_wiring_completeness.py that was needed to
discover the non-conforming filename.
